feat: appointment cards style will help users understand the appointment status
This implementation updates the appointment cards style in order to help users visually understand the appointment status.
It adds also the reply icons on the cards to see the user reply.
It add an informative description of the status inside the summary view and the displayer of the cards.
Finally it enhanced the editor free busy selector and the calendar selector. The free busy will reflect the same color as the selected calendar. All the colors of the free busy status have been updated.